requirement.txt
matplotlib
numpy
opencv-python
rknn-toolkit
ffmpeg-python
flask

